'Project Prioritization Protocol' objective value

Calculate the objective value for a set of solutions to a Project Prioritization Protocol problem.

ppp_objective_value(x, tree, project_column_name, success_column_name,
  solution)

Arguments

x
data.frame or tibble table containing project data. Here, each row should correspond to a different project and columns should contain data that correspond to each project. This object should contain data that denote (i) the name of each project (specified in the argument to project_column_name), (ii) the cost of each project (specified in the argument to cost_column_name), (iii) the probability that each project will succeed if it is funded (specified in the argument to success), and (iv) the enhanced probability that each species will persist if it is funded (specified as the tip labels in the argument to tree). To account for the combined benefits of multiple actions (e.g. baiting and trapping different invasive species in the same area), additional projects should be created that indicate the combined cost and corresponding species' persistence probabilities. Furthermore, this object must have a baseline project, with a zero cost, that represents the probability that each species will persist if no other conservation project is funded.

tree
phylo phylogenetic tree describing the evolutionary history of the species affected by the conservation projects that could potentially be funded. Note that every single species that is affected by the various conservation projects should be represented in this tree.

project_column_name
character name of column that contains the name for each conservation project. Note that the project names must not contain any duplicates or missing values.

success_column_name
character name of column that denotes the probability that each project will succeed if it is funded. This column must have numeric values which lay between zero and one. No missing values are permitted.

solution
tibble object containing the solution data. Here, each row corresponds to a different solution. This object should contain a column for each project---with the column name matching the project name---and the column should contain logical values that indicate if the project is funded (TRUE) or not (FALSE). Although additional columns can also be included, they will be ignored.

Value

numeric vector with an objective value for each solution in the argument to solutions.