Simulate conservation project data

Simulate data to prioritize project funding schemes for species conservation.

ppp_simulate_data(number_species, cost_mean = 100, cost_sd = 5,
  success_min_probability = 0.7, success_max_probability = 0.99,
  funded_min_persistence_probability = 0.5,
  funded_max_persistence_probability = 0.9,
  not_funded_min_persistence_probability = 0.01,
  not_funded_max_persistence_probability = 0.4,
  locked_in_proportion = 0, locked_out_proportion = 0)

Arguments

number_species
numeric number of species.

cost_mean
numeric average cost for the projects. Defaults to 100.

cost_sd
numeric standard deviation in project costs. Defaults to 5.

success_min_probability
numeric minimum probability of the projects succeeding if they are funded. Defaults to 0.7.

success_max_probability
numeric maximum probability of the projects succeeding if they are funded. Defaults to 0.99.

funded_min_persistence_probability
numeric minimum probability of the species persisting if their projects are funded and successful. Defaults to 0.5.

funded_max_persistence_probability
numeric maximum probability of the species persisting if their projects are funded and successful. Defaults to 0.9.

not_funded_min_persistence_probability
numeric minimum probability of the species persisting if their projects are not funded. Defaults to 0.01.

not_funded_max_persistence_probability
numeric maximum probability of the species persisting if their projects are not funded. Defaults to 0.4.

locked_in_proportion
numeric of projects that are locked into the solution. Defaults to 0.

locked_out_proportion
numeric of projects that are locked into the solution. Defaults to 0.

Value

A list object containing the elements:

"project_data"

A tibble containing the data for the conservation projects. It contains the following columns:

"cost"

numeric cost for each project.

"success"

numeric probability of each project succeeding if it is funded.

"locked_in"

logical indicating if certain projects should be locked into the solution.

"locked_out"

logical indicating if certain projects should be locked out of the solution.

"S1" ... "SN"

numeric columns for each species, ranging from "S1" to "SN" where N is the number of species, indicating the enhanced probability that each species will survive if it funded.

tree

phylo phylogenetic tree for the species.

Details

The simulated data set will contain one conservation project for each species and a baseline "do nothing" project to reflect species' persistence when their conservation projects are not funded. The data are simulated as follows:

A conservation project is created for each species.

Cost data for each project are simulated using a normal distribution and the cost_mean and cost_sd arguments.

The probability of each project succeeding is simulated by drawing probabilities from a uniform distribution with the upper and lower bounds set as the success_min_probability and success_max_probability arguments.

The probability of each project succeeding is simulated by drawing probabilities from a uniform distribution with the upper and lower bounds set as the success_min_probability and success_max_probability arguments.

A set proportion of the projects are randomly set to be locked in and out of the solutions using the locked_in_proportion and locked_out_proportion arguments.

The probability of each species persisting is simulated by drawing probabilities from a uniform distribution with the upper and lower bounds set as the funded_min_persistence_probability and funded_max_persistence_probability arguments.

An additional project is created which represents the baseline "do nothing" scenario. The probability of each species persisting when managed under this project is simulated by drawing probabilities from a uniform distribution with the upper and lower bounds set as the not_funded_min_persistence_probability and not_funded_max_persistence_probability arguments.

A phylogenetic tree is simulated for the species using rcoal.